Separate unit and integration tests (#376)
* Separate unit and integration tests and select backward compatibility test suite

This PR separates the test suite into basically two categories: Unit and Integration.

The unit tests do not require a server to run.

Integration tests are also separated into two. Backward compatibility test suite and others.

The tests in the backward compatibility test suite use only the public API and should be safe
to run in any released version of the 4.x client.

The other tests uses some part of the internal API (such as connection manager, invocation service, etc.)
and these might change from version to version. I tried to keep the surface of the
possibly-backward-incompatible integration tests as small as possible.

This PR generally left the tests as they are, changes introduced are either due to test separation or
using public API etc.

* compatibility fixes

* remove backward_compatibility folder

* make unit tests faster

+++ hazelcast/near_cache.py
@@ -42,7 +42,7 @@ class DataRecord(object):
         """Determines whether this record is expired or not.
 
         Args:
-            max_idle_seconds (int): The maximum idle time of record, maximum time after the last access time.
+            max_idle_seconds (float): The maximum idle time of record, maximum time after the last access time.
 
         Returns:
             bool: ``True`` is this record is not expired, ``False`` otherwise.
